folder-gateway-skill/ai-math-service/requirements.txt
fastapi
uvicorn[standard]
httpx
redis
celery
python-dotenv
pydantic
